requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.2
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
entrypoints==0.3
enum34==1.1.6
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
ipaddress==1.0.22
itypes==1.1.0
jeepney==0.4
Jinja2==2.10.1
keyring==19.0.2
keyrings.alt==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycairo==1.18.1
pycparser==2.19
pycrypto==2.6.1
PyGObject==3.32.2
PyJWT==1.7.1
python-dateutil==1.5
pytz==2019.1
pyxdg==0.26
requests==2.22.0
SecretStorage==3.1.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.6.1
whitenoise==4.1.2
